Move unrelated files out of project root
