+++ karma.config.js
@@ -5,9 +5,6 @@ module.exports = function (config) {
   config.set({
     singleRun: false,
     autoWatch: true,
-    // No launchers currently.
-    // visit http://localhost:9876/debug.html
-    // after running `karma start`.
     port: 5555,
     browsers: [],
     files: [ 'tests.webpack.js' ],
@@ -18,5 +15,4 @@ module.exports = function (config) {
     reporters: [ 'progress' ],
     webpack: require("./tools/webpack.config.base.js"),
   });
-  open("http://localhost:5555/debug.html");
 };
