+++ src/assets.tsx
@@ -28,6 +28,7 @@ require("!style!css!sass!./css/index.scss");
 // TODO: Put this into a bundle with file splitting (for caching / performance).
 // As it stands, app will not look right in offline mode because of dependence
 // on these URLs for CSS.
+// TODO: Use webpack CDN loader instead of this mess.
 loadFromCdn(
   "//maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css",
   "css");
